Tree API: No need to set every texture seperately anymore
You only have to give a texture prefix (for example "default_birch" or "default_acacia") and the rest will be done for you!
But that doesn't works for the normal (apple) tree, the jungletree and the needles of the pine tree.

So the code of nine nodes is now reducable to 4-5 lines! I think that is not bad, is it?